Feat. 메뉴 선택 기능 분리 & 게임 일시정지 기능 구현
